Rename match store
